requirements.txt
# Core dependencies
selenium==4.15.2
pandas==1.5.3
openpyxl
python-dotenv==1.0.0

# Database (optionnel)
# sqlalchemy==2.0.23
# pymysql==1.1.0
# psycopg2-binary==2.9.9
pyodbc==5.0.1

# AI
openai==1.3.7

# API REST (optionnel)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Utils
python-dateutil==2.8.2
pyyaml==6.0.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Development
# black==23.11.0
# flake8==6.1.0
# mypy==1.7.1
